refactor: 스텝 공고 상세 페이지 버튼 여백 수정
