+++ app/step/index.tsx
@@ -2,7 +2,7 @@ import { router } from "expo-router";
 import { Pressable, Text } from "react-native";
 
 import CustomSafeAreaView from "@/src/components/layout/CustomSafeAreaView";
-import BackArrorHeader from "@/src/components/ui/BackArrorHeader";
+import BackArrorHeader from "@/src/components/ui/BackArrowHeader";
 
 export default function Step() {
   return (
